+++ src/Exception/MassAssignmentException.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * @author Jared King <j@jaredtking.com>
+ *
+ * @see http://jaredtking.com
+ *
+ * @copyright 2015 Jared King
+ * @license MIT
+ */
+
+namespace Pulsar\Exception;
+
+class MassAssignmentException extends ModelException
+{
+}